product -8
