+++ src/pages/company/styles.module.scss
@@ -8,7 +8,10 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
-
+    .main{
+      width: 80%;
+      margin-left: 10%;
+    }
     p{
       font-size: 20px;
       line-height: 40px;
@@ -21,17 +24,11 @@
     text-align: center;
   }
 }
-.main{
-  width: 80%;
-  margin-left: 10%;
-}
-
 h1{
   font-size: 50px;
   line-height: 60px;
   
 }
-
 h2{
   font-size: 40px;
   line-height: 50px;
@@ -40,7 +37,6 @@ h2{
   text-align: center;
   position: relative;
 }
-
 .trust:after{
   content: "";
   position: absolute;
