+++ src/pages/deployment/plancard/styles.module.scss
@@ -1,99 +1,132 @@
 .PlanCard{
   display: flex;
   align-items: center;
-  width: 100%;
-  max-width: 640px;
-  justify-content: center;
+  width: 48%;
   flex-direction: column;
   background: #FFFFFF;
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
   border-radius: 20px;
   border:  4px transparent solid;
-  margin: 0 40px 150px 40px;
+  margin-bottom: 50px;
   .title{
-    p{
-      width: 344px;
+    width: 100%;
+  text-align: center; 
+    height: 110px;
+    display: flex;
+    align-items: center;
+  font-size: 30px;
+  line-height: 40px;
+  overflow: hidden;
+  img{
+    display: block;
+    width: 40%;
+    max-width: 120px;
+    margin:0 0 0 22px ;
+  }
+  p{
+    display: block;
+    width: 100%;
+    text-align: center;
+    margin: 0;
+  }
+  }
+  .instruction{
+    flex-direction: column;
+    display: flex;
+    text-align: center; 
+    justify-content: center;
+    align-items: center;
+    .instructionmain{
+      font-family: Inter;
+      font-size: 24px;
+      font-weight: 300;
+      line-height: 30px;
+      text-align: left;
+      margin-bottom: 33px;
+      font-size: 24px;
+      width: 70%;
+  
     }
+    .button{
+      width: 80%;
+      height: 64px;
+      font-size: 24px;
+      background: #2580DE;
+      border-radius: 8px;
+    }
+    .functionlist{
+      display: flex;
+      justify-content: center;
+      ul{
+        padding: 0;
+        margin: 0;
+      }
+      li{
+        margin-bottom: 29px;
+        list-style: none;
+        font-size: 20px;
+        font-weight: 400;
+        line-height: 24px;
+        text-align: left;
+        img{
+          margin-right: 8px;
+        }
+      }
+    
+    }
+    
   }
 }
-.PlanCard:hover{
-  border-radius: 20px;
-  border:  4px transparent solid;
+.blue{
   border-image: linear-gradient(207.95deg, #88C3FF -12.8%, #A2EADD 49.13%, #C5EA81 112.8%) 20;
   clip-path: inset(0 round 10px);
   .title{
-    p{
-      width: 344px;
-    }
-    background: linear-gradient(207.95deg, #399BFF -12.8%, #62E7D0 112.78%, #B1EC44 112.8%);
-color: #FFFFFF;
+  background: linear-gradient(207.95deg, #399BFF -12.8%, #62E7D0 112.78%, #B1EC44 112.8%);
+  color: #FFFFFF;
   }
 }
-.title{
-  width: 100%;
-  // background: linear-gradient(207.95deg, #399BFF -12.8%, #62E7D0 112.78%, #B1EC44 112.8%);
-text-align: center; 
-// color: #FFFFFF;
-  height: 110px;
-  font-family: Inter;
-  display: flex;
-  align-items: center;
-font-size: 36px;
-font-weight: 300;
-line-height: 40px;
-padding: 15px 0 0 0 ;
-overflow: hidden;
-img{
-  margin:0 52px 0 22px ;
-}
-p{
-  margin: 0;
-}
-}
-.instruction{
-  flex-direction: column;
-  display: flex;
-  text-align: center; 
-  justify-content: center;
-  align-items: center;
-  margin-top: 65px;
-  .instructionmain{
-    font-family: Inter;
-    font-size: 24px;
-    font-weight: 300;
-    line-height: 30px;
-    text-align: left;
-    margin-bottom: 33px;
-    font-size: 24px;
-    width: 70%;
-    min-height: 280px;
-  }
-}
-.functionlist{
-  display: flex;
-  justify-content: center;
-  ul{
-    padding: 0;
-    margin: 0;
-  }
-  li{
-    margin-bottom: 29px;
-    list-style: none;
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 24px;
-    text-align: left;
-    img{
-      margin-right: 8px;
-    }
-  }
 
+
+
+
+.list{
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  justify-content: space-around;
 }
-.button{
-  margin: 55px 0 78px 0;
-  width: 80%;
-  height: 64px;
-  font-size: 24px;
-  background: #2580DE;
-  border-radius: 8px;
+
+
+
+
+@media screen and (max-width: 996px) {
+  .PlanCard{
+    width: 100%!important;
+  }
+  .title{
+    p{
+      font-size: 30px!important;
+    }
+  }
+  .instruction{
+    p{
+      font-size: 24px!important;
+      line-height: 32px!important;
+    }
+  }
+}
+
+@media screen and (max-width: 450px) {
+  .title{
+    p{
+      font-size: 20px!important;
+    }
+  }
+  .instruction{
+    p{
+      font-size: 20px!important;
+      line-height: 28px!important;
+    }
+  }
+  
 }
