+++ src/pages/deployment/styles.module.scss
@@ -1,24 +1,32 @@
-.bck{
-  position: absolute;
-  top: 60px;
-  left: 0;
+.background{
   width: 100%;
-  height: 700px;
-  background: url("@site/static/img/product/back.png") no-repeat center;
-  z-index: -1;
-}.title{
-  font-family: Inter;
-  font-size: 64px;
-  max-width: 1500px;
-  // font-style: italic;
-  font-weight: 400;
-  letter-spacing: -0.025em;
-  color: #2580DE;
-  margin: 120px auto;
-  text-align: center;
-}.first{
+  background: url("@site/static/img/product/back.png") no-repeat top;
+  background-size: 100% 50%;
+}
+.first{
   display: flex;
   flex-direction: row;
   justify-content: center;
   align-items: center;
+}
+.title{
+  text-align: center;
+  font-size: 50px;
+  line-height: 60px;
+}
+
+@media screen and (max-width: 996px) {
+  .title{
+    font-size: 40px !important;
+    line-height: 48px !important;
+  }
+
+}
+
+@media screen and (max-width: 450px) {
+  .title{
+    font-size: 1.5em !important;
+    line-height: 30px !important;
+  }
+
 }
