+++ src/pages/product/feature/styles.module.scss
@@ -46,8 +46,11 @@
   h1{
     line-height: 48px!important;
   }
+  h2{
+    line-height: 30px!important;
+  }
   p{
-    line-height: 28px!important;
+    line-height: 20px!important;
   }
 }
 @media screen and (max-width: 450px) {
